refactor(engine): drop @tatolab/core dep; engine is fully wire-schema-free
Engine no longer declares `@tatolab/core` in `libs/streamlib-engine/streamlib.yaml`.
The wire vocabulary lives in packages that actually speak it on the wire;
the engine carries bytes through iceoryx2 and operates on raw fields
(string surface IDs, u32 dimensions, Option<i32> layout enumerants, etc.).
Per-package JTD codegen + the serialization layer is the cross-language
contract — engine never needs to know any schema type.

What changed:

- Engine's `streamlib.yaml` keeps only the dep + schemas it actually
  speaks: `@tatolab/escalate` for the polyglot escalate IPC protocol
  the engine itself implements. The seven `@tatolab/core` schemas
  (AudioFrame / ColorInfo / ContentLight / EncodedAudioFrame /
  EncodedVideoFrame / MasteringDisplay / VideoFrame) are gone.

- Engine's test mock processors (TestMockProcessor / TestMockInputOnly /
  TestMockOutputOnly / TestProcessor) now declare `schema: any` on their
  ports — the wildcard sentinel already supported by the processor-schema
  parser. The mocks exercise engine's attribute-macro + graph-compiler
  contracts without coupling to any wire schema.

- Engine's `gpu_context.rs` tests that previously constructed
  `crate::_generated_::VideoFrame` purely to extract `surface_id` /
  `width` / `height` / `texture_layout` and pass them to the raw-field
  API now call the API directly with literals. The VideoFrame was test
  ceremony that obscured what the tests actually exercise — engine's
  raw-field registration and resolution paths.

- `videoframe_texture_layout_serialization_round_trip` was a pure JTD
  serialization test (no engine logic exercised). Moves alongside
  `videoframe_color_metadata_round_trip` into `packages/core/tests/`
  where the wire schema actually lives.

Net result: engine has no `_generated_::tatolab__core::*` types and no
`@tatolab/core` Cargo dep through the codegen pipeline. iceoryx2 carries
bytes between processors; each package generates its own Rust types from
the JTD schemas it declares in its own `streamlib.yaml`; engine routes
those bytes without ever materializing them as typed Rust frames at
its API surface.
